Refactor zustand store and it's name import
